requirements.txt
asgiref==3.2.5
boto3==1.12.31
botocore==1.15.31
cffi==1.14.0
cryptography==2.8
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.4
django-cors-headers==3.2.1
django-environ==0.4.5
django-heroku==0.3.1
django-rest-framework==0.1.0
django-rest-knox==4.1.0
django-s3-storage==0.13.0
django-storages==1.9.1
djangorestframework==3.11.0
docopt==0.6.2
docutils==0.15.2
gitignore==0.0.8
gunicorn==20.0.4
jmespath==0.9.5
psycopg2==2.8.4
pycparser==2.20
pycrypto==2.6.1
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
static3==0.7.0
urllib3==1.25.8
whitenoise==5.0.1
